gst/gstvalue.c: fix for #308599 - fixes deserializing gint64 types
Original commit message from CVS:
* gst/gstvalue.c:
fix for #308599 - fixes deserializing gint64 types
* gst/parse/grammar.y:
add a log line to show what type we deserialize to

+++ gst/gstvalue.c
@@ -979,7 +979,7 @@ gst_value_deserialize_boolean (GValue * dest, const char *s)
  * int *
  *******/
 
-static int
+static long long
 gst_strtoll (const char *s, char **end, int base)
 {
   long long i;
